refactor crate::KdbxEntry -> crate::diff::Entry
